# swagger.yaml
openapi: 3.0.0
info:
  title: Scalable REST API
  version: v1
  description: API documentation for the backend assignment. For full interactive docs, use /docs on the running backend.
paths:
  /api/v1/auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '200':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
  /api/v1/auth/login:
    post:
      summary: Login and get JWT
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
  /api/v1/tasks:
    get:
      summary: Get all tasks for user
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskOut'
    post:
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '200':
          description: Created task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskOut'
components:
  schemas:
    UserCreate:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    UserOut:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        is_admin:
          type: boolean
    Token:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
    TaskCreate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
    TaskOut:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        owner_id:
          type: integer